# Speed up hashing committed tables