--- a/splitgraph/__init__.py
+++ b/splitgraph/__init__.py
@@ -8,4 +8,3 @@ from .config import CONFIG
 from .core.repository import Repository
 from .engine import get_engine, switch_engine
 from .exceptions import SplitGraphException
-from .splitfile import *