--- a/splitgraph/commandline/image_info.py
+++ b/splitgraph/commandline/image_info.py
@@ -12,12 +12,11 @@ from splitgraph.core._common import pretty_size
 from splitgraph.core._drawing import render_tree
 from splitgraph.core.engine import get_current_repositories
 from splitgraph.core.object_manager import ObjectManager
-from splitgraph.core.repository import Repository
-from ._common import image_spec_parser, pluralise
+from ._common import ImageType, pluralise, RepositoryType
 
 
 @click.command(name='log')
-@click.argument('repository', type=Repository.from_schema)
+@click.argument('repository', type=RepositoryType())
 @click.option('-t', '--tree', is_flag=True)
 def log_c(repository, tree):
     """
@@ -43,7 +42,7 @@ def log_c(repository, tree):
 @click.option('-v', '--verbose', default=False, is_flag=True,
               help='Include the actual differences rather than just the total number of updated rows.')
 @click.option('-t', '--table-name', help='Show the differences for a single table.')
-@click.argument('repository', type=Repository.from_schema)
+@click.argument('repository', type=RepositoryType())
 @click.argument('tag_or_hash_1', required=False)
 @click.argument('tag_or_hash_2', required=False)
 def diff_c(verbose, table_name, repository, tag_or_hash_1, tag_or_hash_2):
@@ -130,7 +129,7 @@ def _get_actual_hashes(repository, image_1, image_2):
 
 
 @click.command(name='show')
-@click.argument('image_spec', type=image_spec_parser(default='HEAD'))
+@click.argument('image_spec', type=ImageType(default='HEAD'))
 @click.option('-v', '--verbose', default=False, is_flag=True,
               help='Also show all tables in this image and the objects they map to.')
 def show_c(image_spec, verbose):
@@ -239,7 +238,7 @@ def sql_c(sql, schema, show_all):
 
 
 @click.command(name='status')
-@click.argument('repository', required=False, type=Repository.from_schema)
+@click.argument('repository', required=False, type=RepositoryType())
 def status_c(repository):
     """
     Show the status of the Splitgraph engine. If a repository is passed, show information about