--- a/splitgraph/hooks/external_objects.py
+++ b/splitgraph/hooks/external_objects.py
@@ -57,7 +57,20 @@ def get_external_object_handler(name, handler_params):
         handler_class = _EXTERNAL_OBJECT_HANDLERS[name]
         return handler_class(handler_params)
     except KeyError:
-        raise SplitGraphException("Protocol %s is not supported!" % name)
+        external_handlers = CONFIG.get('external_handlers', {})
+        if name not in external_handlers:
+            raise SplitGraphException("Protocol %s is not supported!" % name)
+        else:
+            handler_class_name = external_handlers[name]
+            index = handler_class_name.rindex('.')
+            try:
+                handler_class = getattr(import_module(handler_class_name[:index]), handler_class_name[index + 1:])
+                register_upload_download_handler(name, handler_class)
+                return handler_class(handler_params)
+            except AttributeError as e:
+                raise SplitGraphException("Error loading external object handler {0}".format(name)) from e
+            except ImportError as e:
+                raise SplitGraphException("Error loading external object handler {0}".format(name)) from e
 
 
 def register_upload_download_handler(name, handler_class):
@@ -67,19 +80,3 @@ def register_upload_download_handler(name, handler_class):
     if name in _EXTERNAL_OBJECT_HANDLERS:
         raise SplitGraphException("Cannot register a protocol handler %s as it already exists!" % name)
     _EXTERNAL_OBJECT_HANDLERS[name] = handler_class
-
-
-def _register_default_handlers():
-    for handler_name, handler_class_name in CONFIG.get('external_handlers', {}).items():
-        index = handler_class_name.rindex('.')
-        try:
-            handler_class = getattr(import_module(handler_class_name[:index]), handler_class_name[index + 1:])
-            register_upload_download_handler(handler_name, handler_class)
-        except AttributeError as e:
-            raise SplitGraphException("Error loading external object handler {0}".format(handler_name)) from e
-        except ImportError as e:
-            raise SplitGraphException("Error loading external object handler {0}".format(handler_name)) from e
-
-
-# Register the default object handlers from the config
-_register_default_handlers()