… test/splitgraph/test_commandline.py
… test/splitgraph/test_object_hashing.py