Hash table fragments after they've been copied over to a temporary location rather than extracting hashes from the whole table with LIMIT.

Improves the time taken to commit a table with 1M rows:

  * original: 88s to create a table, 310s to commit (275s in `calculate_content_hash`);
  * new: 74s to commit (29s in `copy_table`, 40s in `calculate_content_hash`)