@mildbyte 2019-04-14
Extra speedups: commandline entry point.

Was:

```
(splitgraph-3.6.8)  ~/splitgraph $ time sgr > /dev/null

real    0m2.176s
user    0m1.992s
sys     0m0.392s
(splitgraph-3.6.8)  ~/splitgraph $ time python -c "import splitgraph.commandline"

real    0m1.702s
user    0m1.544s
sys     0m0.368s
```

Moved imports of most external libraries to be lazy (Minio, Pandas, Parsimonious grammar compilation). Now:

```
(splitgraph-3.6.8)  ~/splitgraph $ time sgr > /dev/null

real    0m0.776s
user    0m0.656s
sys     0m0.040s
(splitgraph-3.6.8)  ~/splitgraph $ time python -c "import splitgraph.commandline"

real    0m0.376s
user    0m0.268s
sys     0m0.044s
```

A lot of the time in commandline invocation is spent in `pkg_resources`:

```
(splitgraph-3.6.8)  ~/splitgraph $ cat `pyenv which sgr`
#!/home/mildbyte/.pyenv/versions/3.6.8/envs/splitgraph-3.6.8/bin/python3.6
# EASY-INSTALL-ENTRY-SCRIPT: 'splitgraph','console_scripts','sgr'
__requires__ = 'splitgraph'


import re
import sys
from pkg_resources import load_entry_point

if __name__ == '__main__':
    sys.argv[0] = re.sub(r'(-script\.pyw?|\.exe)?$', '', sys.argv[0])
    sys.exit(
        load_entry_point('splitgraph', 'console_scripts', 'sgr')()
    )

(splitgraph-3.6.8)  ~/splitgraph $ time python -c "import pkg_resources"

real    0m0.631s
user    0m0.464s
sys     0m0.084s
```